Update Template Inference for Apollo, Nuxt, Dojo (#15)
Changed template inference for Apollo, Nuxt and Dojo to use the pattern `totalDependencies.some(dep => <dependenciesList>.includes(dep))` to check against a list of package names, which is the equivalent to multiple OR conditionals.

some() will return true if one or more items in the callback returns true, and includes() will return true if an item exists within an array.

For each template that could match against multiple packages, those packages have been placed in their own list, for example:
const nuxt = ["nuxt", "nuxt-edge"]

if (totalDependencies.some(dep => nuxt.includes(dep))) {
  return "nuxt";
}

Because apollo-server has multiple different packages to support different server frameworks (such as Express, Hapi, Koa, etc), a list has been added for those packages so that sandboxes which include them will be properly inferred to have the Apollo template. This is necessary to support importing Apollo server repositories from GitHub. Currently if you use the existing Apollo template and change it's dependencies to use a different variant, such as apollo-server-lambda, the sandbox will automatically switch to the Node template. Upon exporting to GitHub from CodeSandbox and re-importing, the template then defaults to create-react-app, which is not the desired behavior, as the sandbox environment is no longer a server sandbox.

Please see the following sandboxes for reference:

https://codesandbox.io/s/github/Saeris/Apollo-Server-Lambda-CodeSandbox-Template/tree/master

https://codesandbox.io/s/4m9l3o6rw